+++ core/src/test/kotlin/sift/core/api/testdata/set1/ClassWithFields.kt
@@ -1,0 +1,8 @@
+package sift.core.api.testdata.set1
+
+class ClassWithFields {
+    private val field: Foo = Foo()
+
+    class Foo
+}
+